directory-editor.ui, launcher-editor.ui: set text fields to expand horizontally
Currently, Launcher and Directory dialogs are resizable, and when widened, text fields do not expand.

This leaves empty dialog to the side, and does not provide additional space to see text field contents.

To fix this, enable hexpand for the text fields and the grids containing them in these dialogs.